+++ package.json
@@ -17,6 +17,7 @@
     "gh-pages": "^3.2.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
     "tabler-icons-react": "^1.37.0",
     "typescript": "^4.6.2",
