+++ src/views/SettingsPage/index.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+
+// components
+// import Header from '../../components/Header/Header';
+import { Container } from '@mantine/core';
+
+// styling
+
+// types
+
+function SettingsPage() {
+
+  return (
+    <>
+      {/* <Header /> */}
+      <Container>
+        <h1>settings</h1>
+      </Container>
+    </>
+  );
+}
+
+export default SettingsPage;
